use vcpkg instead of brew to handle dependencies (#159)
* - added vcpkg as submodule
- added vcpkg.json
- updated CI files

* - update submodules

* don't install doctest and pkgconfig with brew

* - removed vcpkg submodule
- attempt to use github action caching

* - use lukka action for vcpkg

* - fixed windows build
- updated compat.yml

* - fixed formatting

* - fixed unit-test

* temporary disable caching for old MacOS

* - don't use makefile

* Revert "- don't use makefile"

This reverts commit 4715df0220bada0a9e2d2f86afaf3aa97cf791f9.

* - don't use makefile

* - fixed formatting
- don't include vcpkg.cmake

* - try to use caching for old MAcOS

* - use github virtual env for vcpkg

* - fix CI

* - use VCPKG_INSTALLATION_ROOT

* - attempt to fix path to vcpkg

* - attempt to fix wrong config

* - conditional read of env

* - removed unnecessary step

* - attempt to fix CI

* - removed vcpkg as submodule

* - removed unused variable

* - use correctly enable_testing() to allow ctest find all tests

* - fix windows buld ctest target

* - invoke enable_testing() before first add_directory() to properly build test list
- addressed review comments

+++ lib/hpke/src/rsa.cpp
@@ -31,7 +31,8 @@ RSASignature::generate_key_pair(size_t bits)
   }
 
   // NOLINTNEXTLINE(hicpp-signed-bitwise)
-  if (EVP_PKEY_CTX_set_rsa_keygen_bits(ctx.get(), bits) <= 0) {
+  if (EVP_PKEY_CTX_set_rsa_keygen_bits(ctx.get(), static_cast<int>(bits)) <=
+      0) {
     throw openssl_error();
   }
 
@@ -62,7 +63,8 @@ std::unique_ptr<Signature::PublicKey>
 RSASignature::deserialize(const bytes& enc) const
 {
   const auto* data_ptr = enc.data();
-  auto* pkey = d2i_PublicKey(EVP_PKEY_RSA, nullptr, &data_ptr, enc.size());
+  auto* pkey = d2i_PublicKey(
+    EVP_PKEY_RSA, nullptr, &data_ptr, static_cast<int>(enc.size()));
   if (pkey == nullptr) {
     throw openssl_error();
   }
@@ -87,7 +89,8 @@ std::unique_ptr<Signature::PrivateKey>
 RSASignature::deserialize_private(const bytes& skm) const
 {
   const auto* data_ptr = skm.data();
-  auto* pkey = d2i_PrivateKey(EVP_PKEY_RSA, nullptr, &data_ptr, skm.size());
+  auto* pkey = d2i_PrivateKey(
+    EVP_PKEY_RSA, nullptr, &data_ptr, static_cast<int>(skm.size()));
   if (pkey == nullptr) {
     throw openssl_error();
   }
